Check for existing node, add recipe, lower anchor placement
This adds a check to ensure that the floating anchor is only placed in a node space that is occupied by air. A recipe was added to allow crafting the floating anchor. The anchor placement was lowered to place the block below the player's feet instead of inside of the player.